pyproject.toml
# uv config
[project]
name = "weather-app"
version = "0.1.0"
# description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "djangorestframework",
  "djangorestframework-simplejwt",
  "django-cors-headers",
  "PyJWT",
  # "pytz", # allows accurate and cross platform timezone calculations
  # "pytzwhere",
  "python-dotenv",
  # Reads key-value pairs from a .env file and sets them as env variables
  "psycopg[binary]",
  "requests",
  "requests_cache",
  "retry_requests",
  "openmeteo_requests",
  "pandas",
  "Django>=5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/weather-app"]
